+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "^4.5.2",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
+    "react-datepicker": "^3.3.0",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.3"
   },
