+++ src/components/NewTask/newTask.module.css
@@ -1,0 +1,3 @@
+.datePicker>div, .datePicker input{
+    width: 100%;
+}
